Android project's output for JDT is now bin/classes/
This matches the bin structure of the Ant build system where
the javac output is in bin/classes/ and the Android specific files
are in bin/
Also, this will allow us better control of android output files
as they won't conflict with javac's output.

New projects are created with this setup and existing projects
are converted as they are opened.

Change-Id: I83339906b0d43d7a988a75927c60b664c183d27f