Fix Palette preview image scaling on Windows
When using palette previews with a scaling factor (e.g. the small or
tiny palette modes), the preview images look bad on Windows.  This
appears to be because SWT on Windows has a different set of defaults
for antialiasing and interpolation.

This checkin turns on antialiased high interpolation for image
painting in the palette (when the palette is scaled) to address the
first issue listed in 15071: Layout editor rendering issues in ADT 10

Change-Id: Ie3c619d9105d9d30b1bbe5fb91c195c0f65fee97