Add drop-down menu button to the palette control
Add a menu button to the palette label line, as is done in most
Eclipse windows, which brings up the context menu for the
palette. This makes the palette modes etc a bit more discoverable.

(I played with adding a few dedicated toggle buttons there as well,
but it looked a bit odd because the palette toolbar is a few pixels
higher than the layout actions bar in the canvas, so visually things
looked misaligned; we'll need to address that before adding proper
action buttons in the palette title panel.)

Change-Id: I18906ee7c2ebede96f167925ee27a7064715fa8b