Merge ec970d641caedee9656ef8fcc7cdd60e6812b75d into tools_r10. do not merge.
This is needed since development.git in tools_R10 is master, so the
moved items are duplicated.
Original commit msg:

Prepare for moving platform-specific emulator modules to development.git

We need to move the platform-specific modules to development.git to
avoid some clashes. This patch uses a guard Make variable to ensure that
each module is only built once.

The plan is to simply copy these modules to development.git, then remove
the current ones from sdk.git. The guard variable is here to ensure that
the transition happens smoothly, i.e. if a build contains modules in
both directories, only one of them will be built.

Change-Id: I8d343650a283ecffc149a0bf4b3ec0f769c9ca75