Prepare for moving platform-specific emulator modules to development.git
We need to move the platform-specific modules to development.git to
avoid some clashes. This patch uses a guard Make variable to ensure that
each module is only built once.

The plan is to simply copy these modules to development.git, then remove
the current ones from sdk.git. The guard variable is here to ensure that
the transition happens smoothly, i.e. if a build contains modules in
both directories, only one of them will be built.

Change-Id: I4bc58741a18488f18a4b17e8a1edc4a4a5a93cbc