Add Quickfix and Quick Assistant for creating resources
This changeset adds two related features:

- A "marker resolution" which adds a quickfix to aapt errors for
  nonexistent resources. This means that you can right click on the
  error in the Problems view to create the associated missing
  resource.

- A "quick assistant" which looks for a missing resource on the
  current line and if found adds a fix handler for it, similar to the
  quickfix above.

When a fix is invoked, it will create the new file or value, and open
it in the editor with the relevant value section selected.

The quick assistant can be invoked by the normal Ctrl-1 (Cmd-1). The
quickfix must be invoked from the Problems view; the Java editor seems
to add an extra level of integration with a lightbulb in the editor
margin but we don't get that in XML.

The quickfixes work for all value-based resources, as well as some
file-based ones (in particular, those supported by the New XML File
wizard, which it reuses.)

This changeset also adds unit tests for quickfixes along with a few
infrastructure changes to support it.

Change-Id: I962bcf6c98934685e4d74389469d0903115a75e3