Fix bounds computation for drags from palette to canvas
When you drag over the canvas, the potential drop location is shown
along with an outline of the component as it would appear at that
location. This relies on knowing the size of the dragged component.

When dragging from the palette we use the drag preview image. However,
the bounds were a bit too big because we were using the raw size of
the drag image, and it now includes a drop shadow!  This changeset
adjusts this such that we use the original ViewInfo bounds from the
render rather than the usually larger image bounds.

Change-Id: Ic4b870995e17284fbd8a7ed2f4b4a7ec02f4e2c3