Make edit texts grow in horizontal LinearLayouts using weights
If you drop a text field in a vertical linear layout, then the *width*
of the text field is automatically set to fill.  This uses metadata to
check what the fill preference is of each dropped view.

This changeset extends this scheme to also grow text fields in a
horizontal linear layout. However, instead of using fill_parent on the
layout_width, this uses a layout_weight instead. This means that the
text field will fill, but it will not push subsequent children out of
the way the way a fill_parent would.

This changeset also changes the default orientation of a LinearLayout
that is created in new XML files. (I looked around and noticed most
layouts that have a LinearLayout at the root will use a vertical
LinearLayout.)

Change-Id: I52eec4841e4bac076da655c603235547cfd46b73