Fix AAPT error range
Turns out the AAPT errors always list a line number that corresponds
to the line where the element starts.  If the error occurs in an
attribute that is on a subsequent line, the code to identify the
editor offset failed.

This changeset addresses that, by using a document forward search to
locate the attribute. The quickfix code was also checking for an exact
line number match, so that code is tweaked as well.

Change-Id: Ib648d2e2e7ad3ba082fd3c6a48e7f077b3aebc10