Add go to declaration support for styles and embedded text
This changeset adds support to the Go To Declaration code for styles.
Style strings are a bit different from resource URLs - they can have
an @android: prefix, but they don't specify a resource type.  For
example, go to declaration can now jump to the value in the parent
attribute:

   <style name="BrowserTheme" parent="@android:Theme.Black">

This will show the Theme.Black style definition in the SDK's
styles.xml document.

In addition, the hyperlink resolver can now also link to fragments of
an attribute value. In particular, in the above, if you point anywhere
to the left of the ".", only the @android:Theme range will be
selected. This lets you go to parents, since styles inherit from other
styles by adding to them with dots.

Third, the hyperlink resolver can now link to resource urls that is in
the content part of XML documents (e.g. a text node, not an element
name or attribute value). As an example, you can now jump to the color
defined inside this style element:

    <style name="CustomTheme" parent="android:Theme.Light">
        <item name="android:bg">@color/custom_theme_color</item>
    </style>

Finally, this changeset adds unit/plugin tests for the hyperlink
resolver, which asserts the expected list of matches, and the expected
editor context when the default link is opened.

Change-Id: I73757ec78405f7c711e13387d0bb046f698799f1