Refactoring Quick Assistant
Add another quick assistant to the editor which offers the various
refactoring operations for the current editing context.  If the caret
is in a value declaration, then the "Extract String" refactoring is
offered. Otherwise, if the caret is on an element tag name (or opening
or closing tag), then the element-related visual refactorings are
offered (wrap in, change widget, change layout, extract as include)
provided of course the current editor is a layout editor.

Change-Id: Iee4033d2766d2d4a1157789b81602d37a5175c88