Handle AAPT ranges in value resource files
The earlier fix to add editor column ranges for AAPT errors handled
the types of errors that occur in layout .xml files. In value
definition files there are several additional error messages that can
be parsed to infer the actual error range on the line.  This changeset
adds handling for these additional cases, plus various unit tests.

It also fixes a couple of corner cases in layout files - in
particular, the case where a resource value is missing, and the case
where there are other -valid- uses of the same value earlier on the
same line where the first reference would be highlighted rather than
the one following the property in error.

Change-Id: I02906ada43a884d7518348924715892de9d33311