15736: Eclipse crashes in layout editor when browsing element's properties
This changeset attempts to fix issue 15736. The root bug is an SWT bug
which can cause a JVM crash in some scenarios. However, it seems to be
triggered by code which performs a gradient paint in the property
sheet (which is in fact done to work around another SWT bug...).

Using a gradient is not necessary - a plain background paint will do -
so this changeset simplifies the code to only fill the background.  It
also leaves the selection flag on the event alone such that other
parts of the list (such as category items) are not affected.

Change-Id: Ica32fe1e6f49f1e4154c3ffad9cae9799fcc4723