Auto-zoom views the first time they are opened
Now that we default to the tablet screen size, creating a new layout
usually opens up a view which only shows the top left part of the
design view, and horizontal and vertical scrollbars.

This changeset adds auto-zoom such that the first time a view is
opened, the view also auto-zooms. Note that the	type of zooming
performed is zoom-out-only; it never zooms in, so a small phone layout
will continue to be shown at 100%.  We already perform auto-zoom in
several other scenarios (such as changing orientation).

Change-Id: I1e9a9e8357c952f84b541ea799c529d5dc4b3239