Enforce at most one child on ScrollViews, and add LinearLayouts
This changeset fixes issue

15899: Scroll view can only have one child and editor should enforce
       this.

First, it modifies the view rules for ScrollView and HorizontalView to
refuse to add children to the view if there is already a child
there. It also updates the drawing code to not offer drop previews
other than the target highlighting rectangle.

Second, it modifies the New XML File wizard to insert a vertical
LinearLayout as the new child when a ScrollView or
HorizontalScrollView is created. (This was already done by the view
rules when the scroll views were dragged in from the palette.)

Change-Id: Ide825cbf28b46177983d6c5f2ea8c2848147c711