SDK Manager 2: Refactor window implementation.
The previous patch based the implementation of the
new window on top of the old UpdaterWindowImpl.
Although that works since there are many things in
common, in practice the goal is to get rid of the
old one eventually so it's easier to not add a
dependency that needs to be removed later.

Change-Id: Ic1eac19dcc7e91d6750b7462eb2d28b447acad3a