Improve painting of combined selection and hover
When the mouse is over a rectangle, we highlight this "hover"
rectangle by painting a semitranslucent white rectangle on top of it.

When a view is selected, we highlight it by painting a semitranslucent
blue rectangle over it.

This means that if you move the mouse over the selection, you get both
of these effects added together, which dilutes the underlying selected
widget too much.

This changeset tries to improve this situation by defining a different
visual style to be used for the combination of hover and selection,
where the opacity is much lower for the hover in this case.

This changeset also reduces the existing hover by about 30% opacity.

Change-Id: I63ffe8a9d756dcae29b2009a1a1cd6b9ffb6fbe7