Dynamically generated resources id were not used in id resolution.
Dynamic IDs are generated when requesting an int value from a
(type,name) pair where type==ResourceType.ID and the name doesn't
reference any existing ID (This is possible when IDs are declared inline
and the layout file has not been saved, for instance, so the R class
has not been regenerated yet).

There is a reverse method querying for a (type,name) pair based on
an integer. This methods was not looking for generated id and would
return null instead of the name associated with the generated ID.

Change-Id: I0bc4bc76a5157d56d06c5f4538c4fcc38f1bded7