New application icons for the various SDK tools
This changeset integrates a new set of application icons by Roman
Nurik for the various SDK tools: DDMS, Hierarchyviewer, Traceview, the
SDK manager, and the Eclipse about box icon.

We have new icons with the same visual look for draw9patch and
emulator as well, which are not integrated here. Draw9patch is written
in Swing, so it doesn't have a way to set a dock icon on the Mac
(other than with a startup script modification to pass -Xdock:icon,
which means we also need to bundle the .icns image outside its
.jar). I couldn't find the place the emulator dock image is set so I
will ask around and file a separate CL for that.

Change-Id: Idbb2212b8d0d5f1bd2d41b3576079ff409c7e9c8