Add list view preview
Use the new layoutlib facility to specify adapter view bindings to
assign default listview item layouts to ListViews and
ExpandableListViews. This is also used to provide a preview of
ListViews in the palette.

Note that the list preview is always showing the default list view
items. A later changeset will add the ability to pick a custom layout
to be used for listview rendering.

This changeset also contains a refactoring which pulls the cyclic
dependency validator out into a standalone class; this validator will
be used not just for include drop support but also for the layout
selector when picking layouts to preview in a listview.

Finally, the CalendarView has a new IViewRule which sets its default
width and height to fill parent.

Change-Id: If4fdec8d5439af69a34b029cb36a940fac953f86