Fix phone sort in configuration chooser.
When trying to figure out the best device to use to render a given
layout, we sort all the potential candidates. The sort would
favor lower densities before higher ones.

This new sort gives us flexibility in which density are higher
priority.
The current order is: high, med, xhigh, low.

Change-Id: Ie0d5a583bb850d0a2888e973a9e93cfac27bc4d3