SDK Manager command line install improvements.
1- In no-ui mode, the "update sdk" command was missing
the add-on filter type, e.g.:
$ android update sdk --no-ui --filter add-on

This restores it and adds a unit-test to make sure
the cmd line check is in sync with the array definitions.

2- Adds a new command line option "list sdk" to list
all possible packages that can be found and updated from
the remote sites. The list has indexes which can then
be used with the "update sdk --filter" option to pick
specific packages to install.

Example:

$ android list sdk
Packages available for install: 7
   1- Android SDK Tools, revision 10
   2- Documentation for Android SDK, API 11, revision 1
...
$ android update sdk --no-ui --filter doc,3,7

This will install "all doc packages" as well as the
packages 3 and 7 mentioned in the "list sdk" output.

3- Changed the "--obsolete" flag to list not only obsolete
packages but actually *all* packages. This works around the
issue that the SDK Manager tries, by default, to only show
unknown platforms that are higher than whatever is currently
installed, so there was no way with the UI to install an
older platform (e.g. if you had API 11 installed, it would
not list APIS 3-10 for new installation.)
I'll revisit this behavior later.

SDK Issue: http://code.google.com/p/android/issues/detail?id=15933

Change-Id: I7626257c39602908058eb7359b4c98cc3f54eef3