Usability fix for the layout actions bar
The layout actions bar shows actions of two types:

* Actions which edit attriubutes of the "current layout"; typically
  the parent of the currently selected views.  For example, the
  "orientation" or "baseline" attributes of a LinearLayout.

* Actions which edit the layout parameters of the selected views. For
  example, the "weight" attribute of children in a LinearLayout.

One thing which was missing is adding in layout actions for views that
are children. For example, the TableView now has an "Insert Row"
action.  If you select the table itself, rather than a child within
the table, you would not see the Insert Row action. Similarly, if you
drop a new LinearLayout, you cannot toggle its orientation attribute;
it won't be shown, or if it is within another LinearLayout, you will
see an orientation toggle but it controls the parent, not the newly
selected LinearLayout.

This changeset addresses this by adding a new section of actions on
the right hand side of the actions bar, which contains the layout
actions which apply to the selection, regardless of the parent type.

For example, if you have a LinearLayout containing a TableLayout, and
you have selected the TableLayout, you will first see the LinearLayout
actions, then the LinearLayout layoutparams actions (which will be
applied to the TableLayout), and finally the TableLayout layout
actions (insert and remove row).

This changeset also improves the TableLayout insert row action to
insert the row before the current selected row (if any) rather than
unconditionally appending it to the end. It also selects the table
after creation to make it more obvious where it was added.  The new
ability to select nodes from layout rules is also used in a couple of
other places.

Change-Id: I7cd8f75e61fc916bc75ed5ad156440f0f8cbd786