SDK Manager: Fix detection of emulator in tools local folder.
That's because the constant used did not have
the extension anymore.

Note: this was only an issue on Windows since mac/linux
don't need the .exe extension in emulator.

Also this doesn't fix the missing "emulator.exe" in AOSP,
but digit has the CL for that. So in AOSP one need to manually
copy tools/emulator-arm to tools/emulator for the local tools
package to be correctly recognized.

Change-Id: I9ab254eec5c28c4d006ba4132d6b0c6d87c01e15