swtmenubar library for the SDK.
This little EPL library provides a way for the SDK apps to integrate
with the Mac menu bar -- that is correctly hook into the About and
Preferences menu items.

On other platforms (Windows, Linux), corresponding items are
added to the provided menu.

The library provides both a Carbon and a Cocoa implementation.
However the Cocoa implemented is currently commented out since
we only link with a Carbon-aware version of SWT.jar.

Added a README that explain how to use this.

Change-Id: I4b8457f0269946df056b5dd597c7263da1c4e784