Cache resource resolver and manifest info
This changeset adds caching of the ResourceResolver in the layout
editor such that it can be reused for successive rendering requests as
long as the configuration does not change.

It also adds caching of the manifest icon and label; these are
invalidated whenever the manifest file is updated and saved.

Change-Id: Idbf77fd96fa60a4ac24e06bbecae882829a8b812