Change logging API used by ITaskMonitor.
ITaskMonitor is used by the SDK Manager to display
status information and progress bar for asynchronous
tasks (e.g. fetching sources, downloading and installing
packages).

This changes the way text is logged by the monitor.
There used to be one setResult() method which historically
was designed to report 1 final message (e.g. "Install
completed") but then this was actually used to add
ongoing logging. So in this change the monitor has
3 replacement methods: log, logError and logVerbose,
which gives us more flexibility in controlling what
gets displayed.

As a side effect, this fixes unit-tests from SdkLib
that relied on previous output that changed in a
recent CL.

Change-Id: I0fa41d59db8f5eea478b88208695ef07e246ba30