Fix duplicate dependencies reported by SDK installer.
The issue was simply that when finding dependencies
we have several rules that overlap and we just need
to take the union of what we find. For example an
add-on needs to comply with both the IPlatformDependency
and the IExactApiLevelDependency, which is why it was
erroneously listed twice.

SDK Bug: http://code.google.com/p/android/issues/detail?id=16239

Change-Id: Idb0895a0d2aa063d0a3a7f5d6a2c9ddd6959e0f6